% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintBlockRandomization.R
\name{pblockran}
\alias{pblockran}
\title{Print Block Randomization Function}
\usage{
pblockran(design.list)
}
\arguments{
\item{design.list}{List in which each element is the randomization sequence for a single group}
}
\value{
None
}
\description{
The purpose of this function is to print the output from \code{blockran}, the block
randomization function.
}
\examples{
pblockran(blockran(4, 5, list(Gender = c("Male", "Female"), History = c("None", "Low", "High"))))
}

